🎨📝🔥👕 Add initial implementation of go generate files
