+++ event_dispatcher_gen.go
@@ -1,0 +1,693 @@
+package disgord
+
+// Warning: This file has been automatically generated by generate/events/main.go
+// Do NOT make changes here, instead adapt events.go and event/events.go and run go generate
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"sync"
+
+	"github.com/andersfylling/disgord/event"
+)
+
+// NewDispatch construct a Dispatch object for reacting to web socket events
+// from discord
+func NewDispatch(ws DiscordWebsocket) *Dispatch {
+	dispatcher := &Dispatch{
+		allChan: make(chan interface{}),
+
+		channelCreateChan:            make(chan *ChannelCreate),
+		channelDeleteChan:            make(chan *ChannelDelete),
+		channelPinsUpdateChan:        make(chan *ChannelPinsUpdate),
+		channelUpdateChan:            make(chan *ChannelUpdate),
+		guildBanAddChan:              make(chan *GuildBanAdd),
+		guildBanRemoveChan:           make(chan *GuildBanRemove),
+		guildCreateChan:              make(chan *GuildCreate),
+		guildDeleteChan:              make(chan *GuildDelete),
+		guildEmojisUpdateChan:        make(chan *GuildEmojisUpdate),
+		guildIntegrationsUpdateChan:  make(chan *GuildIntegrationsUpdate),
+		guildMemberAddChan:           make(chan *GuildMemberAdd),
+		guildMemberRemoveChan:        make(chan *GuildMemberRemove),
+		guildMemberUpdateChan:        make(chan *GuildMemberUpdate),
+		guildMembersChunkChan:        make(chan *GuildMembersChunk),
+		guildRoleCreateChan:          make(chan *GuildRoleCreate),
+		guildRoleDeleteChan:          make(chan *GuildRoleDelete),
+		guildRoleUpdateChan:          make(chan *GuildRoleUpdate),
+		guildUpdateChan:              make(chan *GuildUpdate),
+		messageCreateChan:            make(chan *MessageCreate),
+		messageDeleteChan:            make(chan *MessageDelete),
+		messageDeleteBulkChan:        make(chan *MessageDeleteBulk),
+		messageReactionAddChan:       make(chan *MessageReactionAdd),
+		messageReactionRemoveChan:    make(chan *MessageReactionRemove),
+		messageReactionRemoveAllChan: make(chan *MessageReactionRemoveAll),
+		messageUpdateChan:            make(chan *MessageUpdate),
+		presenceUpdateChan:           make(chan *PresenceUpdate),
+		presencesReplaceChan:         make(chan *PresencesReplace),
+		readyChan:                    make(chan *Ready),
+		resumedChan:                  make(chan *Resumed),
+		typingStartChan:              make(chan *TypingStart),
+		userUpdateChan:               make(chan *UserUpdate),
+		voiceServerUpdateChan:        make(chan *VoiceServerUpdate),
+		voiceStateUpdateChan:         make(chan *VoiceStateUpdate),
+		webhooksUpdateChan:           make(chan *WebhooksUpdate),
+
+		ws: ws,
+
+		listeners:      make(map[string][]interface{}),
+		listenOnceOnly: make(map[string][]int),
+
+		shutdown: make(chan struct{}),
+	}
+
+	return dispatcher
+}
+
+// Dispatch holds all the channels and internal state for all registered
+// observers
+type Dispatch struct {
+	allChan chan interface{} // any event
+
+	channelCreateChan            chan *ChannelCreate
+	channelDeleteChan            chan *ChannelDelete
+	channelPinsUpdateChan        chan *ChannelPinsUpdate
+	channelUpdateChan            chan *ChannelUpdate
+	guildBanAddChan              chan *GuildBanAdd
+	guildBanRemoveChan           chan *GuildBanRemove
+	guildCreateChan              chan *GuildCreate
+	guildDeleteChan              chan *GuildDelete
+	guildEmojisUpdateChan        chan *GuildEmojisUpdate
+	guildIntegrationsUpdateChan  chan *GuildIntegrationsUpdate
+	guildMemberAddChan           chan *GuildMemberAdd
+	guildMemberRemoveChan        chan *GuildMemberRemove
+	guildMemberUpdateChan        chan *GuildMemberUpdate
+	guildMembersChunkChan        chan *GuildMembersChunk
+	guildRoleCreateChan          chan *GuildRoleCreate
+	guildRoleDeleteChan          chan *GuildRoleDelete
+	guildRoleUpdateChan          chan *GuildRoleUpdate
+	guildUpdateChan              chan *GuildUpdate
+	messageCreateChan            chan *MessageCreate
+	messageDeleteChan            chan *MessageDelete
+	messageDeleteBulkChan        chan *MessageDeleteBulk
+	messageReactionAddChan       chan *MessageReactionAdd
+	messageReactionRemoveChan    chan *MessageReactionRemove
+	messageReactionRemoveAllChan chan *MessageReactionRemoveAll
+	messageUpdateChan            chan *MessageUpdate
+	presenceUpdateChan           chan *PresenceUpdate
+	presencesReplaceChan         chan *PresencesReplace
+	readyChan                    chan *Ready
+	resumedChan                  chan *Resumed
+	typingStartChan              chan *TypingStart
+	userUpdateChan               chan *UserUpdate
+	voiceServerUpdateChan        chan *VoiceServerUpdate
+	voiceStateUpdateChan         chan *VoiceStateUpdate
+	webhooksUpdateChan           chan *WebhooksUpdate
+
+	ws DiscordWebsocket
+
+	listeners      map[string][]interface{}
+	listenOnceOnly map[string][]int
+
+	shutdown chan struct{}
+
+	listenersLock sync.RWMutex
+}
+
+// EventChan ... TODO
+func (d *Dispatch) EventChan(event string) (channel interface{}, err error) {
+	switch event {
+
+	case EventChannelCreate:
+		channel = d.ChannelCreate()
+	case EventChannelDelete:
+		channel = d.ChannelDelete()
+	case EventChannelPinsUpdate:
+		channel = d.ChannelPinsUpdate()
+	case EventChannelUpdate:
+		channel = d.ChannelUpdate()
+	case EventGuildBanAdd:
+		channel = d.GuildBanAdd()
+	case EventGuildBanRemove:
+		channel = d.GuildBanRemove()
+	case EventGuildCreate:
+		channel = d.GuildCreate()
+	case EventGuildDelete:
+		channel = d.GuildDelete()
+	case EventGuildEmojisUpdate:
+		channel = d.GuildEmojisUpdate()
+	case EventGuildIntegrationsUpdate:
+		channel = d.GuildIntegrationsUpdate()
+	case EventGuildMemberAdd:
+		channel = d.GuildMemberAdd()
+	case EventGuildMemberRemove:
+		channel = d.GuildMemberRemove()
+	case EventGuildMemberUpdate:
+		channel = d.GuildMemberUpdate()
+	case EventGuildMembersChunk:
+		channel = d.GuildMembersChunk()
+	case EventGuildRoleCreate:
+		channel = d.GuildRoleCreate()
+	case EventGuildRoleDelete:
+		channel = d.GuildRoleDelete()
+	case EventGuildRoleUpdate:
+		channel = d.GuildRoleUpdate()
+	case EventGuildUpdate:
+		channel = d.GuildUpdate()
+	case EventMessageCreate:
+		channel = d.MessageCreate()
+	case EventMessageDelete:
+		channel = d.MessageDelete()
+	case EventMessageDeleteBulk:
+		channel = d.MessageDeleteBulk()
+	case EventMessageReactionAdd:
+		channel = d.MessageReactionAdd()
+	case EventMessageReactionRemove:
+		channel = d.MessageReactionRemove()
+	case EventMessageReactionRemoveAll:
+		channel = d.MessageReactionRemoveAll()
+	case EventMessageUpdate:
+		channel = d.MessageUpdate()
+	case EventPresenceUpdate:
+		channel = d.PresenceUpdate()
+	case EventPresencesReplace:
+		channel = d.PresencesReplace()
+	case EventReady:
+		channel = d.Ready()
+	case EventResumed:
+		channel = d.Resumed()
+	case EventTypingStart:
+		channel = d.TypingStart()
+	case EventUserUpdate:
+		channel = d.UserUpdate()
+	case EventVoiceServerUpdate:
+		channel = d.VoiceServerUpdate()
+	case EventVoiceStateUpdate:
+		channel = d.VoiceStateUpdate()
+	case EventWebhooksUpdate:
+		channel = d.WebhooksUpdate()
+	default:
+		err = errors.New("no event channel exists for given event: " + event)
+	}
+
+	return
+}
+
+// alwaysListenToChans makes sure no deadlocks occure
+func (d *Dispatch) alwaysListenToChans() {
+	go func() {
+		stop := false
+		for {
+			select {
+			case <-d.allChan:
+
+			case <-d.channelCreateChan:
+			case <-d.channelDeleteChan:
+			case <-d.channelPinsUpdateChan:
+			case <-d.channelUpdateChan:
+			case <-d.guildBanAddChan:
+			case <-d.guildBanRemoveChan:
+			case <-d.guildCreateChan:
+			case <-d.guildDeleteChan:
+			case <-d.guildEmojisUpdateChan:
+			case <-d.guildIntegrationsUpdateChan:
+			case <-d.guildMemberAddChan:
+			case <-d.guildMemberRemoveChan:
+			case <-d.guildMemberUpdateChan:
+			case <-d.guildMembersChunkChan:
+			case <-d.guildRoleCreateChan:
+			case <-d.guildRoleDeleteChan:
+			case <-d.guildRoleUpdateChan:
+			case <-d.guildUpdateChan:
+			case <-d.messageCreateChan:
+			case <-d.messageDeleteChan:
+			case <-d.messageDeleteBulkChan:
+			case <-d.messageReactionAddChan:
+			case <-d.messageReactionRemoveChan:
+			case <-d.messageReactionRemoveAllChan:
+			case <-d.messageUpdateChan:
+			case <-d.presenceUpdateChan:
+			case <-d.presencesReplaceChan:
+			case <-d.readyChan:
+			case <-d.resumedChan:
+			case <-d.typingStartChan:
+			case <-d.userUpdateChan:
+			case <-d.voiceServerUpdateChan:
+			case <-d.voiceStateUpdateChan:
+			case <-d.webhooksUpdateChan:
+			case <-d.shutdown:
+				stop = true
+			}
+
+			if stop {
+				break
+			}
+		}
+	}()
+}
+
+func (d *Dispatch) triggerChan(ctx context.Context, evtName string, session Session, box interface{}) {
+	prepareBox(evtName, box)
+
+	switch evtName {
+
+	case EventChannelCreate:
+		d.channelCreateChan <- box.(*ChannelCreate)
+	case EventChannelDelete:
+		d.channelDeleteChan <- box.(*ChannelDelete)
+	case EventChannelPinsUpdate:
+		d.channelPinsUpdateChan <- box.(*ChannelPinsUpdate)
+	case EventChannelUpdate:
+		d.channelUpdateChan <- box.(*ChannelUpdate)
+	case EventGuildBanAdd:
+		d.guildBanAddChan <- box.(*GuildBanAdd)
+	case EventGuildBanRemove:
+		d.guildBanRemoveChan <- box.(*GuildBanRemove)
+	case EventGuildCreate:
+		d.guildCreateChan <- box.(*GuildCreate)
+	case EventGuildDelete:
+		d.guildDeleteChan <- box.(*GuildDelete)
+	case EventGuildEmojisUpdate:
+		d.guildEmojisUpdateChan <- box.(*GuildEmojisUpdate)
+	case EventGuildIntegrationsUpdate:
+		d.guildIntegrationsUpdateChan <- box.(*GuildIntegrationsUpdate)
+	case EventGuildMemberAdd:
+		d.guildMemberAddChan <- box.(*GuildMemberAdd)
+	case EventGuildMemberRemove:
+		d.guildMemberRemoveChan <- box.(*GuildMemberRemove)
+	case EventGuildMemberUpdate:
+		d.guildMemberUpdateChan <- box.(*GuildMemberUpdate)
+	case EventGuildMembersChunk:
+		d.guildMembersChunkChan <- box.(*GuildMembersChunk)
+	case EventGuildRoleCreate:
+		d.guildRoleCreateChan <- box.(*GuildRoleCreate)
+	case EventGuildRoleDelete:
+		d.guildRoleDeleteChan <- box.(*GuildRoleDelete)
+	case EventGuildRoleUpdate:
+		d.guildRoleUpdateChan <- box.(*GuildRoleUpdate)
+	case EventGuildUpdate:
+		d.guildUpdateChan <- box.(*GuildUpdate)
+	case EventMessageCreate:
+		d.messageCreateChan <- box.(*MessageCreate)
+	case EventMessageDelete:
+		d.messageDeleteChan <- box.(*MessageDelete)
+	case EventMessageDeleteBulk:
+		d.messageDeleteBulkChan <- box.(*MessageDeleteBulk)
+	case EventMessageReactionAdd:
+		d.messageReactionAddChan <- box.(*MessageReactionAdd)
+	case EventMessageReactionRemove:
+		d.messageReactionRemoveChan <- box.(*MessageReactionRemove)
+	case EventMessageReactionRemoveAll:
+		d.messageReactionRemoveAllChan <- box.(*MessageReactionRemoveAll)
+	case EventMessageUpdate:
+		d.messageUpdateChan <- box.(*MessageUpdate)
+	case EventPresenceUpdate:
+		d.presenceUpdateChan <- box.(*PresenceUpdate)
+	case EventPresencesReplace:
+		d.presencesReplaceChan <- box.(*PresencesReplace)
+	case EventReady:
+		d.readyChan <- box.(*Ready)
+	case EventResumed:
+		d.resumedChan <- box.(*Resumed)
+	case EventTypingStart:
+		d.typingStartChan <- box.(*TypingStart)
+	case EventUserUpdate:
+		d.userUpdateChan <- box.(*UserUpdate)
+	case EventVoiceServerUpdate:
+		d.voiceServerUpdateChan <- box.(*VoiceServerUpdate)
+	case EventVoiceStateUpdate:
+		d.voiceStateUpdateChan <- box.(*VoiceStateUpdate)
+	case EventWebhooksUpdate:
+		d.webhooksUpdateChan <- box.(*WebhooksUpdate)
+	default:
+		fmt.Printf("------\nTODO\nImplement channel for `%s`\n------\n\n", evtName)
+	}
+}
+
+func (d *Dispatch) triggerCallbacks(ctx context.Context, evtName string, session Session, box interface{}) {
+	switch evtName {
+
+	case EventChannelCreate:
+		for _, listener := range d.listeners[EventChannelCreate] {
+			(listener.(ChannelCreateCallback))(session, box.(*ChannelCreate))
+		}
+	case EventChannelDelete:
+		for _, listener := range d.listeners[EventChannelDelete] {
+			(listener.(ChannelDeleteCallback))(session, box.(*ChannelDelete))
+		}
+	case EventChannelPinsUpdate:
+		for _, listener := range d.listeners[EventChannelPinsUpdate] {
+			(listener.(ChannelPinsUpdateCallback))(session, box.(*ChannelPinsUpdate))
+		}
+	case EventChannelUpdate:
+		for _, listener := range d.listeners[EventChannelUpdate] {
+			(listener.(ChannelUpdateCallback))(session, box.(*ChannelUpdate))
+		}
+	case EventGuildBanAdd:
+		for _, listener := range d.listeners[EventGuildBanAdd] {
+			(listener.(GuildBanAddCallback))(session, box.(*GuildBanAdd))
+		}
+	case EventGuildBanRemove:
+		for _, listener := range d.listeners[EventGuildBanRemove] {
+			(listener.(GuildBanRemoveCallback))(session, box.(*GuildBanRemove))
+		}
+	case EventGuildCreate:
+		for _, listener := range d.listeners[EventGuildCreate] {
+			(listener.(GuildCreateCallback))(session, box.(*GuildCreate))
+		}
+	case EventGuildDelete:
+		for _, listener := range d.listeners[EventGuildDelete] {
+			(listener.(GuildDeleteCallback))(session, box.(*GuildDelete))
+		}
+	case EventGuildEmojisUpdate:
+		for _, listener := range d.listeners[EventGuildEmojisUpdate] {
+			(listener.(GuildEmojisUpdateCallback))(session, box.(*GuildEmojisUpdate))
+		}
+	case EventGuildIntegrationsUpdate:
+		for _, listener := range d.listeners[EventGuildIntegrationsUpdate] {
+			(listener.(GuildIntegrationsUpdateCallback))(session, box.(*GuildIntegrationsUpdate))
+		}
+	case EventGuildMemberAdd:
+		for _, listener := range d.listeners[EventGuildMemberAdd] {
+			(listener.(GuildMemberAddCallback))(session, box.(*GuildMemberAdd))
+		}
+	case EventGuildMemberRemove:
+		for _, listener := range d.listeners[EventGuildMemberRemove] {
+			(listener.(GuildMemberRemoveCallback))(session, box.(*GuildMemberRemove))
+		}
+	case EventGuildMemberUpdate:
+		for _, listener := range d.listeners[EventGuildMemberUpdate] {
+			(listener.(GuildMemberUpdateCallback))(session, box.(*GuildMemberUpdate))
+		}
+	case EventGuildMembersChunk:
+		for _, listener := range d.listeners[EventGuildMembersChunk] {
+			(listener.(GuildMembersChunkCallback))(session, box.(*GuildMembersChunk))
+		}
+	case EventGuildRoleCreate:
+		for _, listener := range d.listeners[EventGuildRoleCreate] {
+			(listener.(GuildRoleCreateCallback))(session, box.(*GuildRoleCreate))
+		}
+	case EventGuildRoleDelete:
+		for _, listener := range d.listeners[EventGuildRoleDelete] {
+			(listener.(GuildRoleDeleteCallback))(session, box.(*GuildRoleDelete))
+		}
+	case EventGuildRoleUpdate:
+		for _, listener := range d.listeners[EventGuildRoleUpdate] {
+			(listener.(GuildRoleUpdateCallback))(session, box.(*GuildRoleUpdate))
+		}
+	case EventGuildUpdate:
+		for _, listener := range d.listeners[EventGuildUpdate] {
+			(listener.(GuildUpdateCallback))(session, box.(*GuildUpdate))
+		}
+	case EventMessageCreate:
+		for _, listener := range d.listeners[EventMessageCreate] {
+			(listener.(MessageCreateCallback))(session, box.(*MessageCreate))
+		}
+	case EventMessageDelete:
+		for _, listener := range d.listeners[EventMessageDelete] {
+			(listener.(MessageDeleteCallback))(session, box.(*MessageDelete))
+		}
+	case EventMessageDeleteBulk:
+		for _, listener := range d.listeners[EventMessageDeleteBulk] {
+			(listener.(MessageDeleteBulkCallback))(session, box.(*MessageDeleteBulk))
+		}
+	case EventMessageReactionAdd:
+		for _, listener := range d.listeners[EventMessageReactionAdd] {
+			(listener.(MessageReactionAddCallback))(session, box.(*MessageReactionAdd))
+		}
+	case EventMessageReactionRemove:
+		for _, listener := range d.listeners[EventMessageReactionRemove] {
+			(listener.(MessageReactionRemoveCallback))(session, box.(*MessageReactionRemove))
+		}
+	case EventMessageReactionRemoveAll:
+		for _, listener := range d.listeners[EventMessageReactionRemoveAll] {
+			(listener.(MessageReactionRemoveAllCallback))(session, box.(*MessageReactionRemoveAll))
+		}
+	case EventMessageUpdate:
+		for _, listener := range d.listeners[EventMessageUpdate] {
+			(listener.(MessageUpdateCallback))(session, box.(*MessageUpdate))
+		}
+	case EventPresenceUpdate:
+		for _, listener := range d.listeners[EventPresenceUpdate] {
+			(listener.(PresenceUpdateCallback))(session, box.(*PresenceUpdate))
+		}
+	case EventPresencesReplace:
+		for _, listener := range d.listeners[EventPresencesReplace] {
+			(listener.(PresencesReplaceCallback))(session, box.(*PresencesReplace))
+		}
+	case EventReady:
+		for _, listener := range d.listeners[EventReady] {
+			(listener.(ReadyCallback))(session, box.(*Ready))
+		}
+	case EventResumed:
+		for _, listener := range d.listeners[EventResumed] {
+			(listener.(ResumedCallback))(session, box.(*Resumed))
+		}
+	case EventTypingStart:
+		for _, listener := range d.listeners[EventTypingStart] {
+			(listener.(TypingStartCallback))(session, box.(*TypingStart))
+		}
+	case EventUserUpdate:
+		for _, listener := range d.listeners[EventUserUpdate] {
+			(listener.(UserUpdateCallback))(session, box.(*UserUpdate))
+		}
+	case EventVoiceServerUpdate:
+		for _, listener := range d.listeners[EventVoiceServerUpdate] {
+			(listener.(VoiceServerUpdateCallback))(session, box.(*VoiceServerUpdate))
+		}
+	case EventVoiceStateUpdate:
+		for _, listener := range d.listeners[EventVoiceStateUpdate] {
+			(listener.(VoiceStateUpdateCallback))(session, box.(*VoiceStateUpdate))
+		}
+	case EventWebhooksUpdate:
+		for _, listener := range d.listeners[EventWebhooksUpdate] {
+			(listener.(WebhooksUpdateCallback))(session, box.(*WebhooksUpdate))
+		}
+		//default:
+		//	fmt.Printf("------\nTODO\nImplement callback for `%s`\n------\n\n", evtName)
+	}
+
+	// remove the run only once listeners
+	d.listenersLock.Lock()
+	defer d.listenersLock.Unlock()
+
+	for _, index := range d.listenOnceOnly[evtName] {
+		// https://github.com/golang/go/wiki/SliceTricks#delete-without-preserving-order
+		d.listeners[evtName][index] = d.listeners[evtName][len(d.listeners[evtName])-1]
+		d.listeners[evtName][len(d.listeners[evtName])-1] = nil
+		d.listeners[evtName] = d.listeners[evtName][:len(d.listeners[evtName])-1]
+
+		if len(d.listeners[evtName]) == 0 {
+			// TODO: call removeEvent from socket pkg
+		}
+	}
+
+	// remove the once only register
+	if _, exists := d.listenOnceOnly[evtName]; exists {
+		delete(d.listenOnceOnly, evtName)
+	}
+}
+
+// ChannelCreate gives access to channelCreateChan for ChannelCreate events
+func (d *Dispatch) ChannelCreate() <-chan *ChannelCreate {
+	d.ws.RegisterEvent(event.ChannelCreate)
+	return d.channelCreateChan
+}
+
+// ChannelDelete gives access to channelDeleteChan for ChannelDelete events
+func (d *Dispatch) ChannelDelete() <-chan *ChannelDelete {
+	d.ws.RegisterEvent(event.ChannelDelete)
+	return d.channelDeleteChan
+}
+
+// ChannelPinsUpdate gives access to channelPinsUpdateChan for ChannelPinsUpdate events
+func (d *Dispatch) ChannelPinsUpdate() <-chan *ChannelPinsUpdate {
+	d.ws.RegisterEvent(event.ChannelPinsUpdate)
+	return d.channelPinsUpdateChan
+}
+
+// ChannelUpdate gives access to channelUpdateChan for ChannelUpdate events
+func (d *Dispatch) ChannelUpdate() <-chan *ChannelUpdate {
+	d.ws.RegisterEvent(event.ChannelUpdate)
+	return d.channelUpdateChan
+}
+
+// GuildBanAdd gives access to guildBanAddChan for GuildBanAdd events
+func (d *Dispatch) GuildBanAdd() <-chan *GuildBanAdd {
+	d.ws.RegisterEvent(event.GuildBanAdd)
+	return d.guildBanAddChan
+}
+
+// GuildBanRemove gives access to guildBanRemoveChan for GuildBanRemove events
+func (d *Dispatch) GuildBanRemove() <-chan *GuildBanRemove {
+	d.ws.RegisterEvent(event.GuildBanRemove)
+	return d.guildBanRemoveChan
+}
+
+// GuildCreate gives access to guildCreateChan for GuildCreate events
+func (d *Dispatch) GuildCreate() <-chan *GuildCreate {
+	d.ws.RegisterEvent(event.GuildCreate)
+	return d.guildCreateChan
+}
+
+// GuildDelete gives access to guildDeleteChan for GuildDelete events
+func (d *Dispatch) GuildDelete() <-chan *GuildDelete {
+	d.ws.RegisterEvent(event.GuildDelete)
+	return d.guildDeleteChan
+}
+
+// GuildEmojisUpdate gives access to guildEmojisUpdateChan for GuildEmojisUpdate events
+func (d *Dispatch) GuildEmojisUpdate() <-chan *GuildEmojisUpdate {
+	d.ws.RegisterEvent(event.GuildEmojisUpdate)
+	return d.guildEmojisUpdateChan
+}
+
+// GuildIntegrationsUpdate gives access to guildIntegrationsUpdateChan for GuildIntegrationsUpdate events
+func (d *Dispatch) GuildIntegrationsUpdate() <-chan *GuildIntegrationsUpdate {
+	d.ws.RegisterEvent(event.GuildIntegrationsUpdate)
+	return d.guildIntegrationsUpdateChan
+}
+
+// GuildMemberAdd gives access to guildMemberAddChan for GuildMemberAdd events
+func (d *Dispatch) GuildMemberAdd() <-chan *GuildMemberAdd {
+	d.ws.RegisterEvent(event.GuildMemberAdd)
+	return d.guildMemberAddChan
+}
+
+// GuildMemberRemove gives access to guildMemberRemoveChan for GuildMemberRemove events
+func (d *Dispatch) GuildMemberRemove() <-chan *GuildMemberRemove {
+	d.ws.RegisterEvent(event.GuildMemberRemove)
+	return d.guildMemberRemoveChan
+}
+
+// GuildMemberUpdate gives access to guildMemberUpdateChan for GuildMemberUpdate events
+func (d *Dispatch) GuildMemberUpdate() <-chan *GuildMemberUpdate {
+	d.ws.RegisterEvent(event.GuildMemberUpdate)
+	return d.guildMemberUpdateChan
+}
+
+// GuildMembersChunk gives access to guildMembersChunkChan for GuildMembersChunk events
+func (d *Dispatch) GuildMembersChunk() <-chan *GuildMembersChunk {
+	d.ws.RegisterEvent(event.GuildMembersChunk)
+	return d.guildMembersChunkChan
+}
+
+// GuildRoleCreate gives access to guildRoleCreateChan for GuildRoleCreate events
+func (d *Dispatch) GuildRoleCreate() <-chan *GuildRoleCreate {
+	d.ws.RegisterEvent(event.GuildRoleCreate)
+	return d.guildRoleCreateChan
+}
+
+// GuildRoleDelete gives access to guildRoleDeleteChan for GuildRoleDelete events
+func (d *Dispatch) GuildRoleDelete() <-chan *GuildRoleDelete {
+	d.ws.RegisterEvent(event.GuildRoleDelete)
+	return d.guildRoleDeleteChan
+}
+
+// GuildRoleUpdate gives access to guildRoleUpdateChan for GuildRoleUpdate events
+func (d *Dispatch) GuildRoleUpdate() <-chan *GuildRoleUpdate {
+	d.ws.RegisterEvent(event.GuildRoleUpdate)
+	return d.guildRoleUpdateChan
+}
+
+// GuildUpdate gives access to guildUpdateChan for GuildUpdate events
+func (d *Dispatch) GuildUpdate() <-chan *GuildUpdate {
+	d.ws.RegisterEvent(event.GuildUpdate)
+	return d.guildUpdateChan
+}
+
+// MessageCreate gives access to messageCreateChan for MessageCreate events
+func (d *Dispatch) MessageCreate() <-chan *MessageCreate {
+	d.ws.RegisterEvent(event.MessageCreate)
+	return d.messageCreateChan
+}
+
+// MessageDelete gives access to messageDeleteChan for MessageDelete events
+func (d *Dispatch) MessageDelete() <-chan *MessageDelete {
+	d.ws.RegisterEvent(event.MessageDelete)
+	return d.messageDeleteChan
+}
+
+// MessageDeleteBulk gives access to messageDeleteBulkChan for MessageDeleteBulk events
+func (d *Dispatch) MessageDeleteBulk() <-chan *MessageDeleteBulk {
+	d.ws.RegisterEvent(event.MessageDeleteBulk)
+	return d.messageDeleteBulkChan
+}
+
+// MessageReactionAdd gives access to messageReactionAddChan for MessageReactionAdd events
+func (d *Dispatch) MessageReactionAdd() <-chan *MessageReactionAdd {
+	d.ws.RegisterEvent(event.MessageReactionAdd)
+	return d.messageReactionAddChan
+}
+
+// MessageReactionRemove gives access to messageReactionRemoveChan for MessageReactionRemove events
+func (d *Dispatch) MessageReactionRemove() <-chan *MessageReactionRemove {
+	d.ws.RegisterEvent(event.MessageReactionRemove)
+	return d.messageReactionRemoveChan
+}
+
+// MessageReactionRemoveAll gives access to messageReactionRemoveAllChan for MessageReactionRemoveAll events
+func (d *Dispatch) MessageReactionRemoveAll() <-chan *MessageReactionRemoveAll {
+	d.ws.RegisterEvent(event.MessageReactionRemoveAll)
+	return d.messageReactionRemoveAllChan
+}
+
+// MessageUpdate gives access to messageUpdateChan for MessageUpdate events
+func (d *Dispatch) MessageUpdate() <-chan *MessageUpdate {
+	d.ws.RegisterEvent(event.MessageUpdate)
+	return d.messageUpdateChan
+}
+
+// PresenceUpdate gives access to presenceUpdateChan for PresenceUpdate events
+func (d *Dispatch) PresenceUpdate() <-chan *PresenceUpdate {
+	d.ws.RegisterEvent(event.PresenceUpdate)
+	return d.presenceUpdateChan
+}
+
+// PresencesReplace gives access to presencesReplaceChan for PresencesReplace events
+func (d *Dispatch) PresencesReplace() <-chan *PresencesReplace {
+	d.ws.RegisterEvent(event.PresencesReplace)
+	return d.presencesReplaceChan
+}
+
+// Ready gives access to readyChan for Ready events
+func (d *Dispatch) Ready() <-chan *Ready {
+	d.ws.RegisterEvent(event.Ready)
+	return d.readyChan
+}
+
+// Resumed gives access to resumedChan for Resumed events
+func (d *Dispatch) Resumed() <-chan *Resumed {
+	d.ws.RegisterEvent(event.Resumed)
+	return d.resumedChan
+}
+
+// TypingStart gives access to typingStartChan for TypingStart events
+func (d *Dispatch) TypingStart() <-chan *TypingStart {
+	d.ws.RegisterEvent(event.TypingStart)
+	return d.typingStartChan
+}
+
+// UserUpdate gives access to userUpdateChan for UserUpdate events
+func (d *Dispatch) UserUpdate() <-chan *UserUpdate {
+	d.ws.RegisterEvent(event.UserUpdate)
+	return d.userUpdateChan
+}
+
+// VoiceServerUpdate gives access to voiceServerUpdateChan for VoiceServerUpdate events
+func (d *Dispatch) VoiceServerUpdate() <-chan *VoiceServerUpdate {
+	d.ws.RegisterEvent(event.VoiceServerUpdate)
+	return d.voiceServerUpdateChan
+}
+
+// VoiceStateUpdate gives access to voiceStateUpdateChan for VoiceStateUpdate events
+func (d *Dispatch) VoiceStateUpdate() <-chan *VoiceStateUpdate {
+	d.ws.RegisterEvent(event.VoiceStateUpdate)
+	return d.voiceStateUpdateChan
+}
+
+// WebhooksUpdate gives access to webhooksUpdateChan for WebhooksUpdate events
+func (d *Dispatch) WebhooksUpdate() <-chan *WebhooksUpdate {
+	d.ws.RegisterEvent(event.WebhooksUpdate)
+	return d.webhooksUpdateChan
+}
