+++ generate/events/events.go.tpl
@@ -1,0 +1,22 @@
+package disgord
+
+// Warning: This file has been automatically generated by generate/events/main.go
+// Do NOT make changes here, instead adapt events.go and event/events.go and run go generate
+
+import (
+	"context"
+
+	"github.com/andersfylling/disgord/event"
+)
+
+{{range .}}
+{{if .IsDiscordEvent}}
+{{.RenderDocs}}
+const Event{{.}} = event.{{.}} {{end}}
+
+func (h *{{.}}) registerContext(ctx context.Context) { h.Ctx = ctx }
+
+// {{.}}Callback is triggered in {{.}} events
+type {{.}}Callback = func(session Session, h *{{.}})
+
+// --------------------------- {{end}}
